Added Integral & Derivative term and did some code clean up The integral term is added and also the total integral is calculated by adding the previous total integral with the integral term. Total integral us used to tune the PID controller.
The derivative term is also changed and the derivative term is just error1_dot and error2_dot which is passed in the function argument. Also, our own derivative term is calculated but it is commented because it throws alot of warning when the code with the derived derivative term is run. Need to fix this issue and get the correct derivative term 
All the code is cleaned up. Matrix is used for doing the calculation for both joint at once, instead of doing it two times.